+++ tests/EscaladeTestCase.php
@@ -59,7 +59,38 @@ abstract class EscaladeTestCase extends DbTestCase
         $_SESSION['glpi_currenttime'] = $ctime;
     }
 
-    public function climbWithTimelineButton(\Ticket $ticket, \Group $group, array $options): void
+    public static function climbTicketMethods(array $methods = []): array
+    {
+        $climb_methods = [
+            [
+                'method' => 'climbWithTimelineButton',
+                'itemtype' => \Group::class,
+            ],
+            [
+                'method' => 'climbWithHistoryButton',
+                'itemtype' => \Group::class,
+            ],
+            [
+                'method' => 'climbWithSolvedTicket',
+                'itemtype' => \Group::class,
+            ],
+            [
+                'method' => 'climbWithRejectSolutionTicket',
+                'itemtype' => \Group::class,
+            ],
+            [
+                'method' => 'climbWithAssignMySelfButton',
+                'itemtype' => \User::class,
+            ],
+        ];
+
+        return array_filter($climb_methods, function ($climb_method) use ($methods) {
+            return in_array($climb_method['method'], $methods);
+        });
+
+    }
+
+    public function climbWithTimelineButton(\Ticket $ticket, \Group $group, array $options = []): void
     {
         $options['ticket_details'] = array_merge(
             $options['ticket_details'] ?? [],
@@ -67,7 +98,7 @@ abstract class EscaladeTestCase extends DbTestCase
                 'id' => $ticket->getID(),
             ],
         );
-        $_POST['comment'] = $options['comment'];
+        $_POST['comment'] = $options['comment'] ?? 'Default comment';
         \PluginEscaladeTicket::timelineClimbAction($group->getID(), $ticket->getID(), $options);
         $ticketgroup = new \Group_Ticket();
         $is_escalate = $ticketgroup->getFromDBByCrit([
@@ -99,6 +130,21 @@ abstract class EscaladeTestCase extends DbTestCase
 
     public function climbWithSolvedTicket(\Ticket $ticket, \Group $group, array $solution_options = []): void
     {
+        $config = new \PluginEscaladeConfig();
+        $conf = $config->find();
+        $conf = reset($conf);
+        $config->getFromDB($conf['id']);
+        $this->assertGreaterThan(0, $conf['id']);
+
+        // Update escalade config
+        $this->updateItem(
+            \PluginEscaladeConfig::class,
+            1,
+            [
+                'solve_return_group' => 1,
+            ] + $conf
+        );
+
         $this->createItem(\ITILSolution::class, array_merge([
             'content' => 'Test Solution',
             'itemtype' => $ticket->getType(),
@@ -132,6 +178,27 @@ abstract class EscaladeTestCase extends DbTestCase
         $is_escalate = $ticketgroup->getFromDBByCrit([
             'tickets_id' => $ticket->getID(),
             'groups_id'  => $group->getID(),
+        ]);
+        $this->assertTrue($is_escalate);
+    }
+
+    public function climbWithAssignMySelfButton(\Ticket $ticket, \User $user): void
+    {
+        $this->updateItem(
+            \Ticket::class,
+            $ticket->getID(),
+            [
+                '_itil_assign' => [
+                    '_type' => "user",
+                    'users_id' => $user->getID(),
+                    'use_notification' => 1,
+                ],
+            ],
+        );
+        $ticket_user = new \Ticket_User();
+        $is_escalate = $ticket_user->getFromDBByCrit([
+            'tickets_id' => $ticket->getID(),
+            'users_id'  => $user->getID(),
         ]);
         $this->assertTrue($is_escalate);
     }